[FLINK-37282] Split KafkaWriter into EOS/non-EOS
Split the easy case of non-transactional writer from the transactional writer to simplify reasoning about the state (e.g. which fields are used when).
