+++ flink-connector-kafka/src/main/java/org/apache/flink/connector/kafka/sink/KafkaWriter.java
@@ -25,11 +25,9 @@ import org.apache.flink.api.connector.sink2.Sink;
 import org.apache.flink.connector.base.DeliveryGuarantee;
 import org.apache.flink.connector.kafka.MetricUtil;
 import org.apache.flink.connector.kafka.sink.internal.FlinkKafkaInternalProducer;
-import org.apache.flink.connector.kafka.sink.internal.TransactionalIdFactory;
 import org.apache.flink.metrics.Counter;
 import org.apache.flink.metrics.MetricGroup;
 import org.apache.flink.metrics.groups.SinkWriterMetricGroup;
-import org.apache.flink.runtime.checkpoint.CheckpointIDCounter;
 import org.apache.flink.streaming.connectors.kafka.internals.metrics.KafkaMetricMutableWrapper;
 import org.apache.flink.util.FlinkRuntimeException;
 
@@ -38,7 +36,6 @@ import org.apache.kafka.clients.producer.ProducerRecord;
 import org.apache.kafka.clients.producer.RecordMetadata;
 import org.apache.kafka.common.Metric;
 import org.apache.kafka.common.MetricName;
-import org.apache.kafka.common.errors.ProducerFencedException;
 import org.apache.kafka.common.errors.UnknownProducerIdException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -46,15 +43,11 @@ import org.slf4j.LoggerFactory;
 import javax.annotation.Nullable;
 
 import java.io.IOException;
-import java.util.ArrayDeque;
-import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Collections;
-import java.util.Deque;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.Optional;
 import java.util.Properties;
 import java.util.function.Consumer;
 
@@ -63,8 +56,7 @@ import static org.apache.flink.util.Preconditions.checkNotNull;
 import static org.apache.flink.util.Preconditions.checkState;
 
 /**
- * This class is responsible to write records in a Kafka topic and to handle the different delivery
- * {@link DeliveryGuarantee}s.
+ * This class is responsible to write records in a Kafka topic without transactions.
  *
  * @param <IN> The type of the input elements.
  */
@@ -81,11 +73,10 @@ class KafkaWriter<IN>
     private static final String KAFKA_PRODUCER_METRICS = "producer-metrics";
 
     private final DeliveryGuarantee deliveryGuarantee;
-    private final Properties kafkaProducerConfig;
-    private final String transactionalIdPrefix;
-    private final KafkaRecordSerializationSchema<IN> recordSerializer;
-    private final Callback deliveryCallback;
-    private final KafkaRecordSerializationSchema.KafkaSinkContext kafkaSinkContext;
+    protected final Properties kafkaProducerConfig;
+    protected final KafkaRecordSerializationSchema<IN> recordSerializer;
+    protected final Callback deliveryCallback;
+    protected final KafkaRecordSerializationSchema.KafkaSinkContext kafkaSinkContext;
     private volatile Exception asyncProducerException;
     private final Map<String, KafkaMetricMutableWrapper> previouslyCreatedMetrics = new HashMap<>();
     private final SinkWriterMetricGroup metricGroup;
@@ -98,14 +89,7 @@ class KafkaWriter<IN>
     // Number of outgoing bytes at the latest metric sync
     private long latestOutgoingByteTotal;
     private Metric byteOutMetric;
-    private FlinkKafkaInternalProducer<byte[], byte[]> currentProducer;
-    private final KafkaWriterState kafkaWriterState;
-    // producer pool only used for exactly once
-    private final Deque<FlinkKafkaInternalProducer<byte[], byte[]>> producerPool =
-            new ArrayDeque<>();
-    private long lastCheckpointId;
-
-    private final Deque<AutoCloseable> producerCloseables = new ArrayDeque<>();
+    protected FlinkKafkaInternalProducer<byte[], byte[]> currentProducer;
 
     private boolean closed = false;
     private long lastSync = System.currentTimeMillis();
@@ -119,23 +103,18 @@ class KafkaWriter<IN>
      *
      * @param deliveryGuarantee the Sink's delivery guarantee
      * @param kafkaProducerConfig the properties to configure the {@link FlinkKafkaInternalProducer}
-     * @param transactionalIdPrefix used to create the transactionalIds
      * @param sinkInitContext context to provide information about the runtime environment
      * @param recordSerializer serialize to transform the incoming records to {@link ProducerRecord}
      * @param schemaContext context used to initialize the {@link KafkaRecordSerializationSchema}
-     * @param recoveredStates state from an previous execution which was covered
      */
     KafkaWriter(
             DeliveryGuarantee deliveryGuarantee,
             Properties kafkaProducerConfig,
-            String transactionalIdPrefix,
             Sink.InitContext sinkInitContext,
             KafkaRecordSerializationSchema<IN> recordSerializer,
-            SerializationSchema.InitializationContext schemaContext,
-            Collection<KafkaWriterState> recoveredStates) {
+            SerializationSchema.InitializationContext schemaContext) {
         this.deliveryGuarantee = checkNotNull(deliveryGuarantee, "deliveryGuarantee");
         this.kafkaProducerConfig = checkNotNull(kafkaProducerConfig, "kafkaProducerConfig");
-        this.transactionalIdPrefix = checkNotNull(transactionalIdPrefix, "transactionalIdPrefix");
         this.recordSerializer = checkNotNull(recordSerializer, "recordSerializer");
         checkNotNull(sinkInitContext, "sinkInitContext");
         this.deliveryCallback =
@@ -165,27 +144,12 @@ class KafkaWriter<IN>
             throw new FlinkRuntimeException("Cannot initialize schema.", e);
         }
 
-        this.kafkaWriterState = new KafkaWriterState(transactionalIdPrefix);
-        this.lastCheckpointId =
-                sinkInitContext
-                        .getRestoredCheckpointId()
-                        .orElse(CheckpointIDCounter.INITIAL_CHECKPOINT_ID - 1);
-        if (deliveryGuarantee == DeliveryGuarantee.EXACTLY_ONCE) {
-            abortLingeringTransactions(
-                    checkNotNull(recoveredStates, "recoveredStates"), lastCheckpointId + 1);
-            this.currentProducer = getTransactionalProducer(lastCheckpointId + 1);
-            this.currentProducer.beginTransaction();
-        } else if (deliveryGuarantee == DeliveryGuarantee.AT_LEAST_ONCE
-                || deliveryGuarantee == DeliveryGuarantee.NONE) {
-            this.currentProducer = new FlinkKafkaInternalProducer<>(this.kafkaProducerConfig, null);
-            producerCloseables.add(this.currentProducer);
-            initKafkaMetrics(this.currentProducer);
-        } else {
-            throw new UnsupportedOperationException(
-                    "Unsupported Kafka writer semantic " + this.deliveryGuarantee);
-        }
-
         initFlinkMetrics();
+    }
+
+    public void initialize() {
+        this.currentProducer = new FlinkKafkaInternalProducer<>(this.kafkaProducerConfig);
+        initKafkaMetrics(this.currentProducer);
     }
 
     @Override
@@ -211,40 +175,19 @@ class KafkaWriter<IN>
 
     @Override
     public Collection<KafkaCommittable> prepareCommit() {
-        if (deliveryGuarantee != DeliveryGuarantee.EXACTLY_ONCE) {
-            return Collections.emptyList();
-        }
-
-        // only return a KafkaCommittable if the current transaction has been written some data
-        if (currentProducer.hasRecordsInTransaction()) {
-            final List<KafkaCommittable> committables =
-                    Collections.singletonList(
-                            KafkaCommittable.of(currentProducer, producerPool::add));
-            LOG.debug("Committing {} committables.", committables);
-            return committables;
-        }
-
-        // otherwise, we commit the empty transaction as is (no-op) and just recycle the producer
-        currentProducer.commitTransaction();
-        producerPool.add(currentProducer);
         return Collections.emptyList();
     }
 
     @Override
     public List<KafkaWriterState> snapshotState(long checkpointId) throws IOException {
-        if (deliveryGuarantee == DeliveryGuarantee.EXACTLY_ONCE) {
-            currentProducer = getTransactionalProducer(checkpointId + 1);
-            currentProducer.beginTransaction();
-        }
-        return Collections.singletonList(kafkaWriterState);
+        return Collections.emptyList();
     }
 
     @Override
     public void close() throws Exception {
         closed = true;
         LOG.debug("Closing writer with {}", currentProducer);
-        closeAll(this::abortCurrentProducer, producerPool::clear);
-        closeAll(producerCloseables);
+        closeAll(currentProducer);
         checkState(
                 currentProducer.isClosed(), "Could not close current producer " + currentProducer);
         currentProducer = null;
@@ -253,103 +196,17 @@ class KafkaWriter<IN>
         checkAsyncException();
     }
 
-    private void abortCurrentProducer() {
-        if (currentProducer.isInTransaction()) {
-            try {
-                currentProducer.abortTransaction();
-            } catch (ProducerFencedException e) {
-                LOG.debug(
-                        "Producer {} fenced while aborting", currentProducer.getTransactionalId());
-            }
-        }
-    }
-
-    @VisibleForTesting
-    Deque<FlinkKafkaInternalProducer<byte[], byte[]>> getProducerPool() {
-        return producerPool;
-    }
-
     @VisibleForTesting
     FlinkKafkaInternalProducer<byte[], byte[]> getCurrentProducer() {
         return currentProducer;
     }
 
-    void abortLingeringTransactions(
-            Collection<KafkaWriterState> recoveredStates, long startCheckpointId) {
-        List<String> prefixesToAbort = new ArrayList<>();
-        prefixesToAbort.add(transactionalIdPrefix);
-
-        final Optional<KafkaWriterState> lastStateOpt = recoveredStates.stream().findFirst();
-        if (lastStateOpt.isPresent()) {
-            KafkaWriterState lastState = lastStateOpt.get();
-            if (!lastState.getTransactionalIdPrefix().equals(transactionalIdPrefix)) {
-                prefixesToAbort.add(lastState.getTransactionalIdPrefix());
-                LOG.warn(
-                        "Transactional id prefix from previous execution {} has changed to {}.",
-                        lastState.getTransactionalIdPrefix(),
-                        transactionalIdPrefix);
-            }
-        }
-
-        try (TransactionAborter transactionAborter =
-                new TransactionAborter(
-                        kafkaSinkContext.getParallelInstanceId(),
-                        kafkaSinkContext.getNumberOfParallelInstances(),
-                        this::getOrCreateTransactionalProducer,
-                        producerPool::add)) {
-            transactionAborter.abortLingeringTransactions(prefixesToAbort, startCheckpointId);
-        }
-    }
-
-    /**
-     * For each checkpoint we create new {@link FlinkKafkaInternalProducer} so that new transactions
-     * will not clash with transactions created during previous checkpoints ({@code
-     * producer.initTransactions()} assures that we obtain new producerId and epoch counters).
-     *
-     * <p>Ensures that all transaction ids in between lastCheckpointId and checkpointId are
-     * initialized.
-     */
-    private FlinkKafkaInternalProducer<byte[], byte[]> getTransactionalProducer(long checkpointId) {
-        checkState(
-                checkpointId > lastCheckpointId,
-                "Expected %s > %s",
-                checkpointId,
-                lastCheckpointId);
-        FlinkKafkaInternalProducer<byte[], byte[]> producer = null;
-        // in case checkpoints have been aborted, Flink would create non-consecutive transaction ids
-        // this loop ensures that all gaps are filled with initialized (empty) transactions
-        for (long id = lastCheckpointId + 1; id <= checkpointId; id++) {
-            String transactionalId =
-                    TransactionalIdFactory.buildTransactionalId(
-                            transactionalIdPrefix, kafkaSinkContext.getParallelInstanceId(), id);
-            producer = getOrCreateTransactionalProducer(transactionalId);
-        }
-        this.lastCheckpointId = checkpointId;
-        assert producer != null;
-        LOG.info("Created new transactional producer {}", producer.getTransactionalId());
-        return producer;
-    }
-
-    private FlinkKafkaInternalProducer<byte[], byte[]> getOrCreateTransactionalProducer(
-            String transactionalId) {
-        FlinkKafkaInternalProducer<byte[], byte[]> producer = producerPool.poll();
-        if (producer == null) {
-            producer = new FlinkKafkaInternalProducer<>(kafkaProducerConfig, transactionalId);
-            producerCloseables.add(producer);
-            producer.initTransactions();
-            initKafkaMetrics(producer);
-        } else {
-            producer.initTransactionId(transactionalId);
-        }
-        return producer;
-    }
-
-    private void initFlinkMetrics() {
+    protected void initFlinkMetrics() {
         metricGroup.setCurrentSendTimeGauge(this::computeSendTime);
         registerMetricSync();
     }
 
-    private void initKafkaMetrics(FlinkKafkaInternalProducer<byte[], byte[]> producer) {
+    protected void initKafkaMetrics(FlinkKafkaInternalProducer<byte[], byte[]> producer) {
         byteOutMetric =
                 MetricUtil.getKafkaMetric(
                         producer.metrics(), KAFKA_PRODUCER_METRICS, "outgoing-byte-total");
